+++ src/main/java/uk/gov/hmcts/probate/validator/OriginalWillSignedDateValidationRule.java
@@ -19,6 +19,7 @@ import static uk.gov.hmcts.probate.model.Constants.BUSINESS_ERROR;
 @RequiredArgsConstructor
 public class OriginalWillSignedDateInPastRule implements ValidationRule {
     private static final String ORIGINAL_WILL_SIGNED_DATE_MUST_BE_IN_THE_PAST = "originalWillSignedDateMustBeInThePast";
+    private static final String ORIGINAL_WILL_SIGNED_DATE_MUST_BE_BEFORE_DATE_OF_DEATH = "originalWillSignedDateMustBeBeforeDateOfDeath";
 
     private final BusinessValidationMessageService businessValidationMessageService;
 
@@ -34,9 +35,14 @@ public class OriginalWillSignedDateInPastRule implements ValidationRule {
     private List<String> getErrorCodeForOriginalWillDateDateNotInThePast(CCDData ccdData) {
         List<String> allErrorCodes = new ArrayList<>();
         LocalDate willSignedDate = ccdData.getOriginalWillSignedDate();
-
-        if (willSignedDate != null && willSignedDate.compareTo(LocalDate.now()) >= 0) {
-            allErrorCodes.add(ORIGINAL_WILL_SIGNED_DATE_MUST_BE_IN_THE_PAST);
+        if (willSignedDate != null) {
+            if (willSignedDate.compareTo(LocalDate.now()) >= 0) {
+                allErrorCodes.add(ORIGINAL_WILL_SIGNED_DATE_MUST_BE_IN_THE_PAST);
+            }
+            LocalDate dateOfDeath = ccdData.getDeceasedDateOfDeath();
+            if (dateOfDeath != null && willSignedDate.compareTo(dateOfDeath) >= 0) {
+                allErrorCodes.add(ORIGINAL_WILL_SIGNED_DATE_MUST_BE_BEFORE_DATE_OF_DEATH);
+            }
         }
 
         return allErrorCodes;
